Fix: remove bad files & obsolete commands ci skip:syst (#78)
* fix : remove bad files & add cross-sections

- metadata:remove bad files
- scripts: implement remove bad files in fetch.py script
- plot: fix wrong axis defintion & add new variables definitions
- helpers: update xsections
- wf: add pv info & weight info in BTA_ttbar_wf

* feat: add features for plotting scripts

* fix: add only for standalone condor& plot code

* [run array] ci: customize ci , fix bugs in workflow

- ci: only trigger if changes
- doc,ci: add customize commit message to run ci
- wf: fix dtype bug

* fix : add nested directory for output & fix fetch printout messages

- condor,runner: add nested directory
- runner: add other options for only
- script: change fetch messages

* fix: change default ci, minor bugs ci:skip syst

- helper: fix bug for missing DeepB case
- ci, doc: reoptimize default ci options
- condor: fix copy commands

* fix: Update xsection.py

- remove 13TeV xsection
- Changes made to the xsec in consistent to BTV private google sheeet:
https://docs.google.com/spreadsheets/d/1WUwRG6b6iiPSUr-bDyLjSA4JW8ePdsY3Q55J7gSq4gc/edit

---------

+++ src/BTVNanoCommissioning/workflows/ctag_eDY_valid_sf.py
@@ -268,7 +268,7 @@ class NanoProcessor(processor.ProcessorABC):
                     btagSFs(sel_jet, self.SF_map, weights, "DeepCSVB", syst_wei)
                     btagSFs(sel_jet, self.SF_map, weights, "DeepCSVC", syst_wei)
         else:
-            genflavor = ak.zeros_like(sel_jet.pt)
+            genflavor = ak.zeros_like(sel_jet.pt, dtype=int)
 
         # Systematics information
         if shift_name is None:
